+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [4.0.0](https://github.com/supercharge/framework/compare/v3.20.4...v4.0.0) - 2023-xx-xx
 
+### Added
+- `@supercharge/hashing`
+    - add `createHash` method: create a Node.js `Hash` instance for a given input
+    - add `md5` method: create a Node.js MD5 hash
+    - add `sha256` method: create a Node.js SHA256 hash
+    - add `sha512` method: create a Node.js SHA512 hash
+
 ### Updated
 - bump dependencies
 - `@supercharge/contracts`
@@ -17,6 +24,10 @@
         - the `isMissing(key)` method now determines whether a value for a given `key` is `undefined` (related to `has(key)`, because `isMissing` is doing the opposite of `has`)
         - rename the `add(key, value)` method to `set(key, value)`
         - remove the `add(key, value)` method
+- `@supercharge/hashing`
+    - removed `bcrypt` package from being installed automatically, users must install it explicitely when the hashing driver should use bcrypt
+    - hashing options require a factory function to return the hash driver constructor
+
 
 
 ## [3.20.4](https://github.com/supercharge/framework/compare/v3.20.3...v3.20.4) - 2023-10-15